Add  delay(1) to myStepper::fwd(), after myStepper::rotate_one_step() is called.
This seems to make the stepper run smoother, but perhaps slower. For some reason this delay is not necessary for the stepper to run smoothly when fwd() and rotate_one_step() are not class member functions, like in hardware_tests/test_stepper_HES; in that sketch, the stepper runs smoothly and swiftly without any calls to delay(). Should think about why this is and whether there is some way to get the stepper to run smoothly but swiftly in MultiSens.